requirements.txt
python-telegram-bot
python-dotenv
openai
selenium
telethon
undetected-chromedriver>=3.5.5
webdriver-manager
setuptools
pytz
